Added the ball
Added logic for drawing and moving the ball, including collision
detection and points. Added a 'prevY' field to the Paddle struct to keep
track of whether a paddle is moving and, if so, in what direction.

Also renamed player1 and player2 to the more descriptive lPlayer and
rPlayer, for left and right respectively, and did some other minor
fixes.