Added on-screen scores and changed some configuration
The horizontal margins of the two players' scores aren't quite
symmetrical, but close enough. The code which updates the strings and
SDL_Surfaces for the scores should be placed in a subroutine of its own,
but when I did so the SDL_Surfaces in question wouldn't show up on the
screen for some reason.

Also moved all of the color and control constants into the header file
for ease of configuration, and made most of the geometry constants
relative to each other to simplify changes in resolution.