Decoupled ball movement and scorekeeping
Moved the logic which checks for and handles points into a seperate
function, updateScore(), and made the ball actually go out the edge of
the court when a point is scored rather than calling reset once the ball
has passed the paddle's reach.

For some reason, SDL_FillRect() doesn't allow a Rect.x < -24, which has
resulted in some duplication of code to avoid clipping where it's
unwanted. It would probably be cleaner to seperate the display code from
the underlying data structure, but it's a bit late for MVC now.

It'll do.